UI updates

+++ main.py
@@ -8,7 +8,10 @@ from pyqtgraph.Qt import QtGui
 import pyqtgraph as pg
 import qdarktheme
 import pandas as pd
+from scipy import stats
+import numpy as np
 from utility import Methods
+from scipy import stats
 
 
 class CSVGraphApp(QMainWindow):
@@ -18,10 +21,24 @@ class CSVGraphApp(QMainWindow):
         self.setWindowTitle(_title)
 
         # GUI ----------
-        main = QHBoxLayout()
+        main = QVBoxLayout()
 
-        # Left Panel
-        self.left_panel = QGridLayout()
+        # Header
+        header = QGridLayout()
+        self.header_title = QLabel('Residual Thickness:')
+        self.header_title.setStyleSheet('''QLabel {font-size: 22px; font-weight: bold; color: #FF8040;}''')
+
+        header.setAlignment(Qt.AlignmentFlag.AlignHCenter)
+        header.setContentsMargins(0, 0, 0, 0)
+        header.addWidget(self.header_title, 0, 0)
+
+        w_header_panel = QWidget()
+        w_header_panel.setLayout(header)
+
+        body = QHBoxLayout()
+
+        # Right Panel
+        self.right_panel = QGridLayout()
         self.csv1_title = QLabel(' ')
         self.csv1_title.setStyleSheet('''QLabel {font-size: 16px; font-weight: bold; color: #606060;}''')
         self.table_csv1 = QTableWidget()
@@ -41,6 +58,19 @@ class CSVGraphApp(QMainWindow):
         self.table_csv2.resizeRowsToContents()
         self.table_csv2.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeMode.Stretch)
 
+        self.right_panel.setAlignment(Qt.AlignmentFlag.AlignTop)
+        self.right_panel.setContentsMargins(0, 8, 0, 0)
+
+        self.right_panel.addWidget(self.csv1_title, 0, 0)
+        self.right_panel.addWidget(self.table_csv1, 1, 0)
+        self.right_panel.addWidget(self.csv2_title, 2, 0)
+        self.right_panel.addWidget(self.table_csv2, 3, 0)
+
+        w_right_panel = QWidget()
+        w_right_panel.setLayout(self.right_panel)
+
+        # Left Panel
+        self.left_panel = QGridLayout()
         self.table_widget = QTableWidget(self)
         self.table_widget.setColumnCount(2)
         self.table_widget.setHorizontalHeaderLabels(['Thickness', 'Amplitude'])
@@ -48,47 +78,81 @@ class CSVGraphApp(QMainWindow):
         self.table_widget.resizeColumnsToContents()
         self.table_widget.resizeRowsToContents()
         self.table_widget.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeMode.Stretch)
-
-        # TODO ----
+        plot_configs = QLabel('Plot settings:')
+        plot_configs.setStyleSheet('''QLabel {font-size: 16px; font-weight: bold; color: #606060;}''')
         label_low = QLabel('Low Filter:')
         self.low_filter = QLineEdit()
-        self.low_filter.setMaxLength(3)  # Set maximum length to 3 characters
+        self.low_filter.setMaxLength(2)  # Set maximum length to 4 characters
         self.low_filter.setValidator(QIntValidator())  # Allow only integer input
         self.low_filter.setText('50')
         label_high = QLabel('High Filter:')
         self.high_filter = QLineEdit()
-        self.high_filter.setMaxLength(3)
+        self.high_filter.setMaxLength(2)
         self.high_filter.setValidator(QIntValidator())
         self.high_filter.setText('99')
-        # TODO ----
+        label_bin = QLabel('Mean Interval:')
+        self.bin_filter = QLineEdit()
+        self.bin_filter.setMaxLength(4)
+        self.bin_filter.setValidator(QIntValidator())
+        self.bin_filter.setText('500')
+        label_x_min = QLabel('X min:')
+        self.x_min = QLineEdit()
+        self.x_min.setMaxLength(4)
+        self.x_min.setValidator(QIntValidator())
+        self.x_min.setText('8')
+        label_x_max = QLabel('X max:')
+        self.x_max = QLineEdit()
+        self.x_max.setMaxLength(4)
+        self.x_max.setValidator(QIntValidator())
+        self.x_max.setText('16')
+        label_y_min = QLabel('Y min:')
+        self.y_min = QLineEdit()
+        self.y_min.setMaxLength(8)
+        self.y_min.setValidator(QIntValidator())
+        self.y_min.setText('-11000')
+        label_y_max = QLabel('Y max:')
+        self.y_max = QLineEdit()
+        self.y_max.setMaxLength(8)
+        self.y_max.setValidator(QIntValidator())
+        self.y_max.setText('1000')
 
         self.left_panel.setAlignment(Qt.AlignmentFlag.AlignTop)
-        self.left_panel.setContentsMargins(8, 30, 8, 8)
+        self.left_panel.setContentsMargins(0, 8, 0, 0)
 
-        self.left_panel.addWidget(self.csv1_title, 0, 0, 1, 2)
-        self.left_panel.addWidget(self.table_csv1, 1, 0, 1, 2)
-        self.left_panel.addWidget(self.csv2_title, 2, 0, 1, 2)
-        self.left_panel.addWidget(self.table_csv2, 3, 0, 1, 2)
-        self.left_panel.addWidget(self.table_widget, 4, 0, 1, 2)
-
-        # TODO ----
-        self.left_panel.addWidget(label_low, 5, 0)
-        self.left_panel.addWidget(self.low_filter, 5, 1)
-        self.left_panel.addWidget(label_high, 6, 0)
-        self.left_panel.addWidget(self.high_filter, 6, 1)
-        # TODO ----
+        self.left_panel.addWidget(plot_configs, 0, 0, 1, 3)
+        self.left_panel.addWidget(label_low, 1, 0)
+        self.left_panel.addWidget(self.low_filter, 1, 1)
+        self.left_panel.addWidget(QLabel('%'), 1, 2)
+        self.left_panel.addWidget(label_high, 2, 0)
+        self.left_panel.addWidget(self.high_filter, 2, 1)
+        self.left_panel.addWidget(QLabel('%'), 2, 2)
+        self.left_panel.addWidget(label_bin, 3, 0)
+        self.left_panel.addWidget(self.bin_filter, 3, 1)
+        self.left_panel.addWidget(QLabel('[mm]'), 3, 2)
+        self.left_panel.addWidget(label_x_min, 4, 0)
+        self.left_panel.addWidget(self.x_min, 4, 1)
+        self.left_panel.addWidget(QLabel('[mm]'), 4, 2)
+        self.left_panel.addWidget(label_x_max, 5, 0)
+        self.left_panel.addWidget(self.x_max, 5, 1)
+        self.left_panel.addWidget(QLabel('[mm]'), 5, 2)
+        self.left_panel.addWidget(label_y_min, 6, 0)
+        self.left_panel.addWidget(self.y_min, 6, 1)
+        self.left_panel.addWidget(QLabel('[mm]'), 6, 2)
+        self.left_panel.addWidget(label_y_max, 7, 0)
+        self.left_panel.addWidget(self.y_max, 7, 1)
+        self.left_panel.addWidget(QLabel('[mm]'), 7, 2)
+        self.left_panel.addWidget(self.table_widget, 8, 0, 1, 3)
+        # self.left_panel.setRowStretch(5, 1)
 
         w_left_panel = QWidget()
         w_left_panel.setLayout(self.left_panel)
 
-        main.addWidget(w_left_panel, 1)
-
-        # Right Panel
-        right_panel = QVBoxLayout()
+        # Central Panel
+        central_panel = QVBoxLayout()
         toolbar = self.addToolBar('menu')
-        self.load_csv1 = QAction(QIcon(r'icons/csv-file.png'), 'Load CSV', self)
+        self.load_csv1 = QAction(QIcon(r'icons/csv-file.png'), 'Load CSV with Thickness data', self)
         self.load_csv1.triggered.connect(lambda: self.open_csv(action=1))
-        self.load_csv2 = QAction(QIcon(r'icons/csv-file_2.png'), 'Load CSV', self)
+        self.load_csv2 = QAction(QIcon(r'icons/csv-file_2.png'), 'Load CSV with Amplitud data', self)
         self.load_csv2.triggered.connect(lambda: self.open_csv(action=2))
         self.plot_button = QAction(QIcon(r'icons/plot-graph.png'), 'Plot Data', self)
         self.plot_button.triggered.connect(self.plot_data)
@@ -100,7 +164,7 @@ class CSVGraphApp(QMainWindow):
         toolbar.addAction(self.plot_button)
         toolbar.addAction(self.plot_clear)
         toolbar.setMovable(False)
-        toolbar.setIconSize(QSize(48, 48))
+        toolbar.setIconSize(QSize(32, 32))
         toolbar.setStyleSheet(
             "QToolBar { background: transparent; border: none; height: 64px;}"
             "QToolBar QToolButton { background: transparent; border: none; }"
@@ -108,23 +172,31 @@ class CSVGraphApp(QMainWindow):
         )
 
         self.plot_widget = pg.PlotWidget()
-        self.plot_title = '<span style="font-size: 22px; font-weight: bold; color: #FF8040;">Residual Thickness:</span>'
-        self.plot_widget.setTitle(self.plot_title)
         self.plot_widget.setLabel('left', 'Elevation [mm]')
         self.plot_widget.setLabel('bottom', 'Thickness [mm]')
         self.plot_widget.setLabel('right', '')
         self.plot_widget.setLabel('top', '')
         self.plot_widget.showGrid(True, True)
-        right_panel.setAlignment(Qt.AlignmentFlag.AlignTop)
-        right_panel.setContentsMargins(0, 8, 36, 0)
+        central_panel.setAlignment(Qt.AlignmentFlag.AlignTop)
+        central_panel.setContentsMargins(0, 0, 0, 0)
 
-        right_panel.addWidget(self.plot_widget)
-        right_panel.addWidget(toolbar)
+        central_panel.addWidget(self.plot_widget)
+        central_panel.addWidget(toolbar)
 
-        w_right_panel = QWidget()
-        w_right_panel.setLayout(right_panel)
+        w_central_panel = QWidget()
+        w_central_panel.setLayout(central_panel)
 
-        main.addWidget(w_right_panel, 5)
+        # Body
+        body.addWidget(w_left_panel, 1)
+        body.addWidget(w_central_panel, 5)
+        body.addWidget(w_right_panel, 1)
+
+        w_body_panel = QWidget()
+        w_body_panel.setLayout(body)
+
+        # Main
+        main.addWidget(w_header_panel, 0)
+        main.addWidget(w_body_panel, 1)
 
         self.main_widget = QWidget()
         self.main_widget.setLayout(main)
@@ -189,14 +261,15 @@ class CSVGraphApp(QMainWindow):
             self.table_widget.setRowCount(self.max_rows_to_drop())
             self.buttons_fillers(self.df_csv1, 0)
             self.buttons_fillers(self.df_csv2, 1)
-            self.plot_title = f'<span style="font-size: 22px; font-weight: bold; color: #FF8040;">Residual Thickness: {self.df_csv1_name}</span>'
-            self.plot_widget.setTitle(self.plot_title)
 
     def buttons_fillers(self, _dataframe: pd, _col: int):
         """ generate the buttons to drop tables """
         for row_idx, column in enumerate(_dataframe.columns):
-            drop_button = QPushButton(f'Drop {column}', self)
-            drop_button.clicked.connect(lambda checked, col=column: self.drop_column(col, _dataframe))
+            if column == 'Elevation':
+                drop_button = QPushButton(f'{column}', self)
+            else:
+                drop_button = QPushButton(f'drop {column}', self)
+                drop_button.clicked.connect(lambda checked, col=column: self.drop_column(col, _dataframe))
             self.table_widget.setCellWidget(row_idx, _col, drop_button)
 
     def max_rows_to_drop(self) -> int:
@@ -204,9 +277,10 @@ class CSVGraphApp(QMainWindow):
 
     def drop_column(self, column, _dataframe):
         if column in _dataframe.columns:
-            _dataframe.drop(columns=column, inplace=True, axis=1)
-            self.update_dropping_buttons(self.df_csv1, self.df_csv2)
-            # print(_dataframe)
+            if column != 'Elevation':
+                _dataframe.drop(columns=column, inplace=True, axis=1)
+                self.update_dropping_buttons(self.df_csv1, self.df_csv2)
+                # print(_dataframe)
 
     def update_dropping_buttons(self, _csv1: pd, _csv2: pd):
         self.table_widget.clearContents()
@@ -223,6 +297,7 @@ class CSVGraphApp(QMainWindow):
             return
 
         self.plot_widget.clear()
+        self.header_title.setText(f'Residual Thickness: {self.df_csv1_name}')
         y = self.df_csv1.Elevation
 
         filtered_thickness = self.methods.apply_filter(self.df_csv1.copy(),
@@ -231,18 +306,38 @@ class CSVGraphApp(QMainWindow):
                                                        high_filter=float(self.high_filter.text()),
                                                        saturation=100)
 
+        # self.plot_widget.plot(x_1, y, pen=pg.mkPen(QColor(249, 127, 63), width=2))
+
         try:
             for name, series in filtered_thickness.items():
-                print(name)
+                color = self.methods.generate_color()
                 if name != 'Elevation':
                     x_1 = filtered_thickness[name]
-                    # self.plot_widget.plot(x_1, y, pen=pg.mkPen(QColor(249, 127, 63), width=2))
-                    color = pg.mkBrush(self.methods.generate_color())
-                    scatter_plot = pg.ScatterPlotItem(size=6, pen=pg.mkPen(None), brush=color)
+
+                    average_by_interval = self.methods.average_by_intervals(_x=x_1,
+                                                                            _y=y,
+                                                                            interval=int(self.bin_filter.text()),
+                                                                            min_elevation=-10000)
+
+                    scatter_plot = pg.ScatterPlotItem(size=2, pen=pg.mkPen(None), brush=pg.mkBrush(color+[65]))
                     scatter_plot.setData(x=x_1, y=y)
+
                     self.plot_widget.addItem(scatter_plot)
 
+                    scatter_average = pg.ScatterPlotItem(size=12, pen=pg.mkPen(None), brush=pg.mkBrush(color+[220]))
+                    scatter_average.setData(x=average_by_interval.x, y=average_by_interval.elevation_bin)
+
+                    self.plot_widget.addItem(scatter_average)
+
+                    self.plot_widget.plotItem.vb.setLimits(xMin=int(self.x_min.text()),
+                                                           xMax=int(self.x_max.text()),
+                                                           yMin=int(self.y_min.text()),
+                                                           yMax=int(self.y_max.text()))
+
         except AttributeError as e:
+            print(e)
+
+        except ValueError as e:
             print(e)
 
     def clear_plot(self) -> None:
